chore(Keybinds): Move defaults to common for easy usage in state
