+++ ui/src/utils/keyboard.rs
@@ -1,83 +1,13 @@
+use common::state::settings::GlobalShortcut;
 use common::state::State;
 use dioxus::prelude::*;
 use dioxus_desktop::use_global_shortcut;
 use dioxus_desktop::wry::application::keyboard::ModifiersState;
 
-#[derive(Eq, PartialEq, Hash, Debug, Clone)]
-pub enum GlobalShortcut {
-    ToggleMute,
-    ToggleDeafen,
-    IncreaseFontSize,
-    DecreaseFontSize,
-}
-
-#[derive(Eq, PartialEq, Debug)]
-pub struct Shortcut {
-    keys: Vec<KeyCode>,             // Keys required
-    modifiers: Vec<ModifiersState>, // Modifier keys required
-    system_shortcut: bool, // Determines if the shortcut should work system-wide i.e. even when uplink is not in focus
-}
-
-impl From<(Vec<KeyCode>, Vec<ModifiersState>, bool)> for Shortcut {
-    fn from(shortcut_tup: (Vec<KeyCode>, Vec<ModifiersState>, bool)) -> Self {
-        Shortcut {
-            keys: shortcut_tup.0,
-            modifiers: shortcut_tup.1,
-            system_shortcut: shortcut_tup.2,
-        }
-    }
-}
-
 #[derive(Props)]
 pub struct Props<'a> {
     on_global_shortcut: EventHandler<'a, GlobalShortcut>,
     // TODO: overrides: Vec<(String, String)> allow for overriding the default bindings
-}
-
-pub fn get_default_keybinds() -> Vec<(GlobalShortcut, Shortcut)> {
-    let alt_or_command = if cfg!(target_os = "macos") {
-        // SUPER is command on mac
-        ModifiersState::SUPER
-    } else {
-        ModifiersState::ALT
-    };
-    Vec::from([
-        // To avoid multi-key conflicts, when using a shortcut that uses multiple `KeyCode` values, it's best to use the `ALT` modifier by default.
-        (
-            GlobalShortcut::IncreaseFontSize,
-            Shortcut::from((
-                // TODO(KeyCode::Add):We need to treat this carefully, keyboard doesn't identify Add as + properly
-                // And as EqualSign, not works + from Numpad
-                vec![KeyCode::EqualSign],
-                vec![ModifiersState::CONTROL, ModifiersState::SHIFT],
-                false,
-            )),
-        ),
-        (
-            GlobalShortcut::DecreaseFontSize,
-            Shortcut::from((
-                vec![KeyCode::Subtract],
-                vec![ModifiersState::CONTROL, ModifiersState::SHIFT],
-                false,
-            )),
-        ),
-        (
-            GlobalShortcut::ToggleMute,
-            Shortcut::from((
-                vec![KeyCode::M],
-                vec![alt_or_command, ModifiersState::SHIFT],
-                true,
-            )),
-        ),
-        (
-            GlobalShortcut::ToggleDeafen,
-            Shortcut::from((
-                vec![KeyCode::D],
-                vec![alt_or_command, ModifiersState::SHIFT],
-                true,
-            )),
-        ),
-    ])
 }
 
 #[allow(non_snake_case)]
@@ -87,7 +17,7 @@ pub fn KeyboardShortcuts<'a>(cx: Scope<'a, Props>) -> Element<'a> {
     }
 
     let state = use_shared_state::<State>(cx)?;
-    let keybinds = get_default_keybinds();
+    let keybinds = common::state::default_keybinds::get_default_keybinds();
 
     cx.render(rsx! {
         for (global_shortcut, shortcut) in keybinds {
